requirements-txt.txt
streamlit
torch
pillow
python-dotenv
google-generativeai
diffusers
transformers
accelerate
